#401: Refactored `RoutingInfo` to be directly under auth, as this can be used by multiple Authentication methods (e.g. ApplicationAccess as used by the Participant Portal)
